Refactor bot and rename package
