+++ foremoney/constants.py
