+++ foremoney/states.py
@@ -1,0 +1,26 @@
+# Conversation states for ForeMoney bot
+(
+    FROM_TYPE,
+    FROM_GROUP,
+    FROM_ACCOUNT,
+    TO_TYPE,
+    TO_GROUP,
+    TO_ACCOUNT,
+    AMOUNT,
+    SUMMARY,
+    ADD_ACCOUNT_NAME,
+    TX_LIST,
+    TX_DETAILS,
+    TX_EDIT_AMOUNT,
+    SETTINGS_MENU,
+    DASHBOARD_ACCOUNTS,
+    AG_TYPE_SELECT,
+    AG_GROUPS,
+    AG_GROUP_MENU,
+    AG_GROUP_RENAME,
+    AG_ADD_GROUP_NAME,
+    AG_ACCOUNTS,
+    AG_ADD_ACCOUNT_NAME,
+    ACCOUNT_MENU,
+    ACCOUNT_RENAME,
+) = range(23)
